AI-description/AI-image-description/requirements.txt
transformers
torch
Pillow
flask
deep-translator
